moves classes from domain to prototype package
