+++ dnd-bot/dnd_bot/logic/prototype/entity.py
